pyproject.toml
[project]
name = "mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "hydra-core>=1.3.2",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "mlflow>=3.10.0",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "wandb>=0.25.0",
]

[dependency-groups]
dev = [
    "loguru>=0.7.3",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]
